requirements.txt
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
requests==2.11.1
Werkzeug==0.11.11
datadog==0.13.0
gunicorn==19.6.0
numpy==1.14.2
scipy==1.0.0
matplotlib==1.5.3
leveldb==0.193
protobuf==3.5.2.post1
python-gflags==3.0.7
PyYAML==3.12
Pillow==5.0.0
six==1.11.0
grpcio==1.10.0
mock==2.0.0
tensorflow==1.6.0
webargs==1.6.0
opencv-python==3.3.0.10
graphviz==0.8.1
dlib==19.7.0
h5py==2.7.1
keras==2.1.5
pycocotools==2.0
scikit-image==0.12.3


